+++ src/core/Gate_MIDI.cpp
@@ -235,12 +235,12 @@ struct Gate_MIDIWidget : ModuleWidget {
 
         menu->addChild(new MenuSeparator);
 
-        menu->addChild(
-            createBoolPtrMenuItem("Velocity mode", "", &module->velocityMode));
+		menu->addChild(createBoolPtrMenuItem("Velocity from gate amplitude", "", &module->velocityMode));
 
-        menu->addChild(
-            createMenuItem("Panic", "", [=]() { module->midiOutput.panic(); }));
-    }
+		menu->addChild(createMenuItem("Reset MIDI (Panic)", "",
+			[=]() {module->midiOutput.panic();}
+		));
+	}
 };
 
 Model* modelGate_MIDI = createModel<Gate_MIDI, Gate_MIDIWidget>("CV-Gate");