+++ src/core/MIDI_Gate.cpp
@@ -22,18 +22,27 @@ struct MIDI_Gate : Module {
 
 	midi::InputQueue midiInput;
 
-	/** [cell][channel] */
+	/** True when a cell's gate is held. [cell][channel] */
 	bool gates[16][16];
-	/** [cell][channel] */
-	float gateTimes[16][16];
-	/** [cell][channel] */
+	/** Last velocity value of cell. [cell][channel] */
 	uint8_t velocities[16][16];
+	/** Triggered when a cell's note is played. [cell][channel] */
+	dsp::PulseGenerator trigPulses[16][16];
 	/** Cell ID in learn mode, or -1 if none. */
 	int learningId;
 	/** [cell] */
 	int8_t learnedNotes[16];
-	bool velocityMode;
+
+	// Settings
+
+	enum VelocityMode {
+		FIXED_MODE,
+		VELOCITY_MODE,
+		AFTERTOUCH_MODE,
+	};
+	VelocityMode velocityMode;
 	bool mpeMode;
+	bool trigMode;
 
 	MIDI_Gate() {
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
@@ -50,17 +59,19 @@ struct MIDI_Gate : Module {
 			}
 		}
 		learningId = -1;
-		panic();
 		midiInput.reset();
-		velocityMode = false;
+		velocityMode = FIXED_MODE;
 		mpeMode = false;
+		trigMode = false;
+		panic();
 	}
 
 	void panic() {
 		for (int i = 0; i < 16; i++) {
 			for (int c = 0; c < 16; c++) {
 				gates[i][c] = false;
-				gateTimes[i][c] = 0.f;
+				velocities[i][c] = 127;
+				trigPulses[i][c].reset();
 			}
 		}
 	}
@@ -74,18 +85,15 @@ struct MIDI_Gate : Module {
 		int channels = mpeMode ? 16 : 1;
 
 		for (int i = 0; i < 16; i++) {
-			outputs[GATE_OUTPUTS + i].setChannels(channels);
 			for (int c = 0; c < channels; c++) {
-				// Make sure all pulses last longer than 1ms
-				if (gates[i][c] || gateTimes[i][c] > 0.f) {
-					float velocity = velocityMode ? (velocities[i][c] / 127.f) : 1.f;
-					outputs[GATE_OUTPUTS + i].setVoltage(velocity * 10.f, c);
-					gateTimes[i][c] -= args.sampleTime;
-				}
-				else {
-					outputs[GATE_OUTPUTS + i].setVoltage(0.f, c);
-				}
+				bool pulse = trigPulses[i][c].process(args.sampleTime);
+				bool gate = pulse || (!trigMode && gates[i][c]);
+				float velocity = gate ? 1.f : 0.f;
+				if (velocityMode != FIXED_MODE)
+					velocity *= velocities[i][c] / 127.f;
+				outputs[GATE_OUTPUTS + i].setVoltage(velocity * 10.f, c);
 			}
+			outputs[GATE_OUTPUTS + i].setChannels(channels);
 		}
 	}
 
@@ -97,12 +105,36 @@ struct MIDI_Gate : Module {
 			} break;
 			// note on
 			case 0x9: {
-				if (msg.getValue() > 0) {
-					pressNote(msg.getChannel(), msg.getNote(), msg.getValue());
+				uint8_t velocity = msg.getValue();
+				if (velocity > 0) {
+					pressNote(msg.getChannel(), msg.getNote(), velocity);
 				}
 				else {
-					// Many stupid keyboards send a "note on" command with 0 velocity to mean "note release"
+					// Note-on event with velocity 0 is an alternative for note-off event.
 					releaseNote(msg.getChannel(), msg.getNote());
+				}
+			} break;
+			// polyphonic pressure/aftertouch
+			case 0xa: {
+				if (velocityMode == AFTERTOUCH_MODE) {
+					uint8_t c = mpeMode ? msg.getChannel() : 0;
+					uint8_t note = msg.getNote();
+					uint8_t velocity = msg.getValue();
+					for (int i = 0; i < 16; i++) {
+						if (learnedNotes[i] == note) {
+							velocities[i][c] = velocity;
+						}
+					}
+				}
+			} break;
+			// channel pressure/aftertouch
+			case 0xd: {
+				if (velocityMode == AFTERTOUCH_MODE) {
+					uint8_t c = mpeMode ? msg.getChannel() : 0;
+					uint8_t velocity = msg.getNote();
+					for (int i = 0; i < 16; i++) {
+						velocities[i][c] = velocity;
+					}
 				}
 			} break;
 			default: break;
@@ -120,8 +152,9 @@ struct MIDI_Gate : Module {
 		for (int i = 0; i < 16; i++) {
 			if (learnedNotes[i] == note) {
 				gates[i][c] = true;
-				gateTimes[i][c] = 1e-3f;
-				velocities[i][c] = vel;
+				if (velocityMode == VELOCITY_MODE)
+					velocities[i][c] = vel;
+				trigPulses[i][c].trigger(1e-3f);
 			}
 		}
 	}
@@ -152,16 +185,18 @@ struct MIDI_Gate : Module {
 
 		json_t* notesJ = json_array();
 		for (int i = 0; i < 16; i++) {
-			json_t* noteJ = json_integer(learnedNotes[i]);
-			json_array_append_new(notesJ, noteJ);
+			json_array_append_new(notesJ, json_integer(learnedNotes[i]));
 		}
 		json_object_set_new(rootJ, "notes", notesJ);
 
-		json_object_set_new(rootJ, "velocity", json_boolean(velocityMode));
+		json_object_set_new(rootJ, "velocity", json_integer(velocityMode));
+
+		json_object_set_new(rootJ, "mpeMode", json_boolean(mpeMode));
+
+		json_object_set_new(rootJ, "trigMode", json_boolean(trigMode));
 
 		json_object_set_new(rootJ, "midi", midiInput.toJson());
 
-		json_object_set_new(rootJ, "mpeMode", json_boolean(mpeMode));
 		return rootJ;
 	}
 
@@ -176,16 +211,23 @@ struct MIDI_Gate : Module {
 		}
 
 		json_t* velocityJ = json_object_get(rootJ, "velocity");
-		if (velocityJ)
-			velocityMode = json_boolean_value(velocityJ);
-
-		json_t* midiJ = json_object_get(rootJ, "midi");
-		if (midiJ)
-			midiInput.fromJson(midiJ);
+		// Boolean in Rack <=v2.6.4
+		if (json_is_true(velocityJ))
+			velocityMode = VELOCITY_MODE;
+		else if (json_is_integer(velocityJ))
+			velocityMode = VelocityMode(json_integer_value(velocityJ));
 
 		json_t* mpeModeJ = json_object_get(rootJ, "mpeMode");
 		if (mpeModeJ)
 			mpeMode = json_boolean_value(mpeModeJ);
+
+		json_t* trigModeJ = json_object_get(rootJ, "trigMode");
+		if (trigModeJ)
+			trigMode = json_boolean_value(trigModeJ);
+
+		json_t* midiJ = json_object_get(rootJ, "midi");
+		if (midiJ)
+			midiInput.fromJson(midiJ);
 	}
 };
 
@@ -252,12 +294,26 @@ struct MIDI_GateWidget : ModuleWidget {
         menu->addChild(new MenuSeparator);
 
         menu->addChild(
-            createBoolPtrMenuItem("Velocity mode", "", &module->velocityMode));
+            createIndexPtrSubmenuItem("Output type", {
+			"Gates",
+			"Triggers",
+		}, &module->trigMode));
+
+		menu->addChild(createIndexPtrSubmenuItem("Output amplitude", {
+			"10V",
+			"Velocity",
+			"Aftertouch",
+		}, &module->velocityMode));
 
         menu->addChild(createBoolPtrMenuItem("MPE mode", "", &module->mpeMode));
 
-        menu->addChild(createMenuItem("Panic", "", [=]() { module->panic(); }));
-    }
+		menu->addChild(new MenuSeparator);
+
+		menu->addChild(createMenuItem("Reset MIDI (Panic)", "", [=]() {
+			module->panic();
+		}));
+	}
+
 };
 
 // Use legacy slug for compatibility