+++ src/patch.cpp
@@ -252,7 +252,6 @@ void Manager::saveAutosave() {
 
 	json_dumpf(rootJ, file, JSON_INDENT(2));
 	std::fclose(file);
-	system::remove(patchPath);
 	system::rename(tmpPath, patchPath);
 }
 
@@ -460,7 +459,7 @@ void Manager::pushRecentPath(std::string path) {
 	// Add path to top of recent patches
 	recent.push_front(path);
 	// Limit recent patches size
-	recent.resize(std::min((int) recent.size(), 10));
+	recent.resize(std::min((int) recent.size(), 20));
 }